Update in the cov_naive() function
The cov_naive() function is calculating the variance-covariance matrix and to calculate the same the mathematical formula consist of the summation term divided by (N-1). But here the cov_naive() function was getting divided by N which was producing a wrong answer when I used the code. After making the change in the function as per the mathematical formula I got the right answer.